+++ src/gwt/src/org/rstudio/core/client/files/filedialog/FileDialogStyles.java
@@ -39,4 +39,6 @@ public interface FileDialogStyles extends CssResource
    String columnName();
    String columnSize();
    String columnDate();
+
+   String fadeWrapper();
 }